Renaming evaluation_port.py to evaluation.py
